Implemented learn buttons in the mapping dialog.
git-svn-id: https://midi-shape-shifter.googlecode.com/svn/trunk@149 4e31dc51-3177-9f86-f460-8117e2103d03

+++ MidiShapeShifter/MidiShapeShifter/Mss/IMssMsgRange.cs
@@ -8,13 +8,13 @@ namespace MidiShapeShifter.Mss
         string GetData1RangeStr(IFactory_MssMsgInfo msgInfoFactory);
         string GetData2RangeStr(IFactory_MssMsgInfo msgInfoFactory);
         bool Equals(object o);
-        void InitPublicMembers(MssMsgType msgType, int data1, int data2);
-        void InitPublicMembers(MssMsgType msgType, int data1RangeBottom, int data1RangeTop, int data2RangeBottom, int data2RangeTop);
+        void InitPublicMembers(MssMsgType msgType, double data1, double data2);
+        void InitPublicMembers(MssMsgType msgType, double data1RangeBottom, double data1RangeTop, double data2RangeBottom, double data2RangeTop);
         bool MsgIsInRange(MssMsg mssMsg);
         MssMsgType MsgType { get; set; }
-        int Data1RangeBottom { get; set; }
-        int Data1RangeTop { get; set; }
-        int Data2RangeBottom { get; set; }
-        int Data2RangeTop { get; set; }
+        double Data1RangeBottom { get; set; }
+        double Data1RangeTop { get; set; }
+        double Data2RangeBottom { get; set; }
+        double Data2RangeTop { get; set; }
     }
 }
